+++ src/pySUMOQt/Widget/GraphWidget.py
@@ -183,11 +183,11 @@ class GraphWidget(RWWidget, Ui_Form):
         painter.setRenderHint(QPainter.Antialiasing)
         self.graphicsView.render(painter)
         
-    def _zoomIn_(self):
+    def zoomIn(self):
         val = self.doubleSpinBox.value() + 0.10
         self.doubleSpinBox.setValue(val)
         
-    def _zoomOut_(self):
+    def zoomOut(self):
         val = self.doubleSpinBox.value() - 0.10
         self.doubleSpinBox.setValue(val)